Up codes

+++ basic/c/ex_scanf03.c
@@ -6,6 +6,12 @@
 // Please enter the third integer: 3
 // Average: 3
 
+//test
+// Please enter the first integer: 3
+// Please enter the second integer: 3
+// Please enter the third integer: 4
+// Average: 3.000000
+
 int main(int argc, char const *argv[])
 {
     int num1, num2 , num3;
@@ -15,7 +21,7 @@ int main(int argc, char const *argv[])
     scanf("%d", &num2);
     printf("Please enter the third integer: ");
     scanf("%d", &num3);
-    int average = (num1 + num2 +num3) / 3;
-    printf("Average: %d\n", average);
+    double average = (num1 + num2 +num3) / 3.;
+    printf("Average: %f\n", average);
     return 0;
 }
